Change package and tidy up config
