+++ orbit-compose-multiplatform/src/androidUnitTest/kotlin/org/orbitmvi/orbit/compose/multiplatform/RobolectricTest.android.kt
@@ -1,0 +1,8 @@
+package org.orbitmvi.orbit.compose.multiplatform
+
+import org.junit.runner.RunWith
+import org.robolectric.RobolectricTestRunner
+
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+@RunWith(RobolectricTestRunner::class)
+public actual abstract class RobolectricTest actual constructor()
