--- orbit-compose-multiplatform/src/commonTest/kotlin/org/orbitmvi/orbit/compose/MockLifecycleOwner.kt
@@ -1,41 +1,0 @@
-/*
- * Copyright 2025 Mikołaj Leszczyński & Appmattus Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.orbitmvi.orbit.compose
-
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.LifecycleRegistry
-
-internal class MockLifecycleOwner : LifecycleOwner {
-    private val registry = LifecycleRegistry(this)
-
-    var currentState: Lifecycle.State
-        get() = registry.currentState
-        set(value) {
-            registry.currentState = value
-        }
-
-    val hasObservers: Boolean
-        get() = registry.observerCount > 0
-
-    fun dispatchEvent(event: Lifecycle.Event) {
-        registry.handleLifecycleEvent(event)
-    }
-
-    override val lifecycle: Lifecycle
-        get() = registry
-}
