+++ orbit-compose-multiplatform/src/commonTest/kotlin/org/orbitmvi/orbit/compose/multiplatform/RobolectricTest.kt
@@ -1,0 +1,4 @@
+package org.orbitmvi.orbit.compose.multiplatform
+
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+public expect abstract class RobolectricTest()
