+++ orbit-compose-multiplatform/src/nativeTest/kotlin/org/orbitmvi/orbit/compose/multiplatform/RobolectricTest.native.kt
@@ -1,0 +1,4 @@
+package org.orbitmvi.orbit.compose.multiplatform
+
+@Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
+public actual abstract class RobolectricTest actual constructor()
